+++ RELEASES.md
@@ -1,8 +1,8 @@
 # RELEASES
 
-## VERSION: 1.0.0 (2024/12)
+**VERSION:** `1.0.0 (2024/12)`
 
-### Addition:
+## Addition:
 - CError - Structure for dealing with data.
 - CHashMap - Hashtable implementation.
 - CHashSet - Hashset implementation.
@@ -13,8 +13,19 @@
 - CString - A standardized, easy to use, string implementation.
 - CVector - Dynamic array implementation.
 
-### Modification:
+## Modification:
 None
 
-### Removal/Deprecation:
+## Removal/Deprecation:
+None
+
+**VERSION:** `1.0.1 (2024/12)`
+
+## Addition:
+- CLinkedList - Provide a implementation for singly/doubly linked lists, where the user can specify which to use.
+
+## Modification:
+- Replaced int64_t of co-routines where signed integers are not required. This includes changes in CHashMap, CHashSet, CString, CVector, Operators. This *MAY* **BREAK** ABI.
+
+## Removal/Deprecation:
 None